More annotations
git-svn-id: https://dslabpc10.epfl.ch/svn/src/s2e/trunk@15934 f94e2f5b-dfd8-4992-a279-46519bc7c46e

+++ qemu/s2e/S2EExecutionState.h
@@ -143,7 +143,7 @@ public:
     void setPc(uint64_t pc);
     void setSp(uint64_t sp);
 
-    bool getReturnAddress(uint32_t *retAddr);
+    bool getReturnAddress(uint64_t *retAddr);
     bool bypassFunction(unsigned paramCount);
     void undoCallAndJumpToSymbolic();
